Restructuring and Godep
New Folder Structure -> only main.go in root folder 

Database classes in Database Folder

Godep now performed, should run on server now
