+++ Database/user.go
@@ -1,4 +1,4 @@
-package main
+package Database
 
 import (
 	"github.com/graphql-go/graphql"
